refactor: remove DbCatalog newtype wrapper
