+++ server/src/db/catalog.rs
@@ -1,4 +1,5 @@
 //! This module contains the implementation of the InfluxDB IOx Metadata catalog
+use std::any::Any;
 use std::{
     collections::{btree_map::Entry, BTreeMap},
     sync::Arc,
@@ -7,10 +8,20 @@ use std::{
 use parking_lot::RwLock;
 use snafu::{OptionExt, Snafu};
 
+use arrow_deps::datafusion::{catalog::schema::SchemaProvider, datasource::TableProvider};
+use data_types::error::ErrorLogger;
+use internal_types::selection::Selection;
+use partition::Partition;
+use query::{
+    exec::stringset::StringSet,
+    provider::{self, ProviderBuilder},
+    PartitionChunk,
+};
+
+use super::DBChunk;
+
 pub mod chunk;
 pub mod partition;
-
-use partition::Partition;
 
 #[derive(Debug, Snafu)]
 pub enum Error {
@@ -70,22 +81,27 @@ impl Catalog {
         }
     }
 
-    // List all partitions in this database
+    /// List all partitions in this database
     pub fn partitions(&self) -> impl Iterator<Item = Arc<RwLock<Partition>>> {
         let partitions = self.partitions.read();
         partitions.values().cloned().collect::<Vec<_>>().into_iter()
     }
 
-    // Get a specific partition by name, returning `None` if there is no such
-    // partition
+    /// Get a specific partition by name, returning `None` if there is no such
+    /// partition
     pub fn partition(&self, partition_key: impl AsRef<str>) -> Option<Arc<RwLock<Partition>>> {
         let partition_key = partition_key.as_ref();
         let partitions = self.partitions.read();
         partitions.get(partition_key).cloned()
     }
 
-    // Create a new partition in the catalog and return a reference to
-    // it. Returns an error if the partition already exists
+    /// List all partition keys in this database
+    pub fn partition_keys(&self) -> Vec<String> {
+        self.partitions.read().keys().cloned().collect()
+    }
+
+    /// Create a new partition in the catalog and return a reference to
+    /// it. Returns an error if the partition already exists
     pub fn create_partition(
         &self,
         partition_key: impl Into<String>,
@@ -116,6 +132,72 @@ impl Catalog {
             .get(partition_key)
             .cloned()
             .context(UnknownPartition { partition_key })
+    }
+
+    /// Return the chunks for specific partition
+    pub fn chunks(&self, partition_key: &str) -> Vec<Arc<DBChunk>> {
+        let partition = match self.partition(partition_key) {
+            Some(partition) => partition,
+            None => return vec![],
+        };
+
+        let partition = partition.read();
+
+        partition
+            .chunks()
+            .map(|chunk| {
+                let chunk = chunk.read();
+                DBChunk::snapshot(&chunk)
+            })
+            .collect()
+    }
+}
+
+impl SchemaProvider for Catalog {
+    fn as_any(&self) -> &dyn Any {
+        self as &dyn Any
+    }
+
+    fn table_names(&self) -> Vec<String> {
+        let mut names = StringSet::new();
+
+        self.partitions().for_each(|partition| {
+            let partition = partition.read();
+            partition.chunks().for_each(|chunk| {
+                let chunk = chunk.read();
+                let db_chunk = DBChunk::snapshot(&chunk);
+                db_chunk.all_table_names(&mut names);
+            })
+        });
+
+        names.into_iter().collect::<Vec<_>>()
+    }
+
+    fn table(&self, table_name: &str) -> Option<Arc<dyn TableProvider>> {
+        let mut builder = ProviderBuilder::new(table_name);
+        for partition_key in self.partition_keys() {
+            for chunk in self.chunks(&partition_key) {
+                if chunk.has_table(table_name) {
+                    // This should only fail if the table doesn't exist which isn't possible
+                    let schema = chunk
+                        .table_schema(table_name, Selection::All)
+                        .expect("cannot fail");
+
+                    // This is unfortunate - a table with incompatible chunks ceases to
+                    // be visible to the query engine
+                    builder = builder
+                        .add_chunk(chunk, schema)
+                        .log_if_error("Adding chunks to table")
+                        .ok()?
+                }
+            }
+        }
+
+        match builder.build() {
+            Ok(provider) => Some(Arc::new(provider)),
+            Err(provider::Error::InternalNoChunks { .. }) => None,
+            Err(e) => panic!("unexpected error: {:?}", e),
+        }
     }
 }
